pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tokenpulse"
version = "0.1.0"
description = "macOS menu bar app to track AI token usage — Claude, ChatGPT, Gemini"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["claude", "chatgpt", "gemini", "token", "usage", "macos", "menubar"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Topic :: Utilities",
]
dependencies = [
    "rumps>=0.4.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
tokenpulse = "tokenpulse.app:run"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/tokenpulse"
Issues = "https://github.com/YOUR_USERNAME/tokenpulse/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["tokenpulse*"]
